# worker/docs/swagger.yaml
info:
  contact: {}
  license: {}
paths:
  /addworker/:
    post:
      parameters:
      - description: raft ID
        in: formData
        name: id
        required: true
        type: integer
      - description: Group ID
        in: formData
        name: gid
        required: true
        type: integer
      - description: remote cluster addr
        in: formData
        name: joinCluster
        type: string
      responses:
        "200":
          description: OK
          schema:
            type: body
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: add new worker
  /del/{gid}/{oid}:
    delete:
      parameters:
      - description: Object ID
        in: path
        name: oid
        required: true
        type: integer
      - description: Group ID
        in: path
        name: gid
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
      summary: Delete an object
  /get/{gid}/{id}:
    get:
      parameters:
      - description: Object ID
        in: path
        name: oid
        required: true
        type: integer
      - description: Group ID
        in: path
        name: gid
        required: true
        type: integer
      produces:
      - application/octet-stream
      - image/png
      - image/jpeg
      - image/gif
      - text/plain
      responses:
        "200":
          description: OK
          schema:
            type: body
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Get an object
  /list/:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              type: integer
            type: array
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: List all gid
  /put/{gid}/:
    post:
      consumes:
      - multipart/form-data
      parameters:
      - description: this is a test file
        in: formData
        name: file
        required: true
        type: file
      - description: Group ID
        in: path
        name: gid
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
      summary: Put an object
swagger: "2.0"